Move timeout_encoding from core/ext/transport/chttp2 to core/lib/transport
